+++ CHANGELOG.md
@@ -39,6 +39,8 @@ All notable changes to this project will be documented in this file
 * changed function name from `vehicle` to `vehicleName` in vehicle module
 * changed function name from `timezone` to `timezoneRandom` in date module
 * deleted function `commonFileType` from `System` module, use `system.fileType` instead
+* changed function name from `number` (by format) to `phoneNumberByFormat` in phone module
+* changed function name from `number` (by country) to `phoneNumberByCountry` in phone module
 
 ### Features